+++ module/src/hooks/use-language-switcher-is-active.tsx
@@ -1,8 +1,8 @@
-
-import React, { ReactNode, useState, useEffect} from 'react';
-import { useRouter } from 'next/router';
-import i18n from '../index';
-import { I18N } from '../types';
+import { useState, useEffect } from "react";
+import { useRouter } from "next/router";
+import i18n from "../index";
+import { I18N } from "../types";
+import { LanguageDataStore } from "../enums/languageDataStore";
 
 /**
  * Returns a boolean react-state indicating if the current selected language equals the one passed to the hook.
@@ -10,23 +10,40 @@ import { I18N } from '../types';
  * @returns boolean react-state
  */
 export default function useLanguageSwitcherIsActive(currentLang: string) {
-	let i18nObj: I18N;
-	i18nObj = i18n() as I18N;
-	const defaultLang: string = i18nObj.defaultLang;
-	const router = useRouter();
-	const [isActive, setIsActive] = useState<boolean>(false);
-	useEffect( () => {
-		let current = false;
-		if (!router.query || !router.query.lang) {
-			current = defaultLang === currentLang;
-		}
-		else {
-			current = router.query.lang === currentLang;
-		}
-		setIsActive(current);
-	},[currentLang, defaultLang, router.query]);
-	return { isActive } as const;
-	// return [isActive] as const;
+  const [isActive, setIsActive] = useState(false);
+  const i18nObj = i18n() as I18N;
+  const router = useRouter();
+  const defaultLang = i18nObj.defaultLang;
+  const languageDataStore = i18nObj.languageDataStore;
+
+  useEffect(() => {
+    if (languageDataStore === LanguageDataStore.QUERY) {
+      let current;
+
+      if (!router.query || !router.query.lang) {
+        current = defaultLang === currentLang;
+      } else {
+        current = router.query?.lang === currentLang;
+      }
+
+      setIsActive(current);
+    }
+  }, [currentLang, defaultLang, router.query, languageDataStore]);
+
+  useEffect(() => {
+    if (languageDataStore === LanguageDataStore.LOCAL_STORAGE) {
+      const localStorageLanguage = window.localStorage.getItem(
+        "next-export-i18n-lang"
+      );
+      let current = defaultLang === currentLang;
+
+      if (localStorageLanguage) {
+        current = localStorageLanguage === currentLang;
+      }
+
+      setIsActive(current);
+    }
+  }, [currentLang, defaultLang, languageDataStore]);
+
+  return { isActive } as const;
 }
-
-