+++ module/src/types/index.tsx
@@ -1,7 +1,10 @@
+import { LanguageDataStore } from "../enums/languageDataStore";
+
 export type Dictionary = { [key: string]: string | Dictionary };
 
 export type I18N = {
   translations: { [language: string]: Dictionary };
   defaultLang: string;
   useBrowserDefault: boolean;
+  languageDataStore?: LanguageDataStore;
 };