+++ CHANGELOG.md
@@ -1,10 +1,19 @@
-# 2.0.0
 
-- React v18
-- Next v12
+# 2.2.0
+### Fix
+- Deletes unused variables: https://github.com/martinkr/next-export-i18n/pull/37
+- Fixes handling of query params for useLanguageQuery using dynamic routes and pages with additional query params: https://github.com/martinkr/next-export-i18n/pull/60
+
+### Features
+- Adds support for links with locale parameter: https://github.com/martinkr/next-export-i18n/pull/52
+- Adds support to store selected language in Local Storage https://github.com/martinkr/next-export-i18n/pull/56
+# 2.0.0
+### Features
+- Support for React v18
+- Support for Next v12
 
 # 1.4.2
-
+### Fix
 - <https://github.com/martinkr/next-export-i18n/issues/24>
 
 # 1.4.1